refactor: move portable shared data and device id in a root base trait to reflect core libraries structure
